+++ packages/app/test.js
@@ -1,0 +1,16 @@
+const { spawn } = require("child_process");
+const ls = spawn("npm", ["-v"], {
+    shell: true,
+});
+
+ls.stdout.on("data", (data) => {
+    console.log(`stdout: ${data}`);
+});
+
+ls.stderr.on("data", (data) => {
+    console.error(`stderr: ${data}`);
+});
+
+ls.on("close", (code) => {
+    console.log(`child process exited with code ${code}`);
+});