+++ packages/web/src/store/index.ts
@@ -1,0 +1,7 @@
+import { reactive } from "vue";
+import Store from "electron-store";
+
+export const store = reactive({
+    directories: [],
+    tasks: [],
+});