+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 3.0.0-beta (2022-02-09)
+# 3.0.0-beta (2022-02-13)
 
 
 ### Bug Fixes
@@ -8,9 +8,12 @@
 
 ### Features
 
+* **app and web:** 添加文件夹管理，添加右键菜单，初始化设置和关于页面 ([24930ad](https://github.com/enncy/online-course-script/commit/24930ad1e32508e227fb36b7008fb2981438cbb7))
 * **app and web:** add elctron-builder in app , init  page view in web ([fe60df6](https://github.com/enncy/online-course-script/commit/fe60df65dfeed1607ab89a941fc6b6eb627132fc))
 * **init:** init project ([0675ac6](https://github.com/enncy/online-course-script/commit/0675ac6a631e8946a52e7e4e655b28faee8248d4))
 * **packages:** init packages : web app scripts ([24e5386](https://github.com/enncy/online-course-script/commit/24e5386ec86dec33cc696fda6b5956785e2c1359))
+* **script:** add cx and zhs login ([8d69e16](https://github.com/enncy/online-course-script/commit/8d69e166fd73cf2ae2d700aa772b78159efbdeaa))
+* **scripts:** add script package ([6f03fb4](https://github.com/enncy/online-course-script/commit/6f03fb4d3545f02f8ccfeb14fa7b5d8169c15348))
+* **test and cx login:** update tests README.md and add new login way of cx : phone-code-login ([ab46a1d](https://github.com/enncy/online-course-script/commit/ab46a1df98c73351518847a629ddaeb7f51a2d4d))
 
-
-
+